refactored out individual regions into components
